+++ CHANGES.md
@@ -1,3 +1,5 @@
 
+  - add [fold4] and [fold8]
+
 # 0.1.0
   - first release